+++ workspace/solutions/ecmascript/src/s0/s7/s6/s8/s_0768_test.ts
@@ -2,7 +2,7 @@ import test from "bun:test";
 
 import { t } from "@/utils/test";
 
-t(
+await t(
     test,
     import.meta.dir,
 );