+++ workspace/solutions/rust/src/s2/s1/s0/s4/mod.rs
@@ -1,0 +1,6 @@
+pub mod s_2104_00;
+pub mod s_2104_01;
+pub mod s_2104_02;
+pub mod s_2104_test;
+
+pub type TSolution = fn(Vec<i32>) -> i64;